Fix special handling of `UnionAll`s
`UnionAll`s that come from concrete struct definitions can't be
subtyped, but they themselves also aren't `isconcretetype`. This
lead to a false positive in error handling, which is fixed in this
commit by making sure to only reject abstract types (which `UnionAll`s
from an abstract type definition also are).